HSEARCH-3510 Prepare the documentation module for the addition of more tests
