+++ documentation/src/test/resources/hibernate-test-elasticsearch.properties
@@ -3,4 +3,6 @@ hibernate.search.backends.backendName.hosts ${test.elasticsearch.host.url}
 hibernate.search.backends.backendName.username ${test.elasticsearch.host.username}
 hibernate.search.backends.backendName.password ${test.elasticsearch.host.password}
 hibernate.search.backends.backendName.log.json_pretty_printing true
+hibernate.search.backends.backendName.index_defaults.lifecycle.strategy drop-and-create-and-drop
+hibernate.search.backends.backendName.index_defaults.lifecycle.minimal_required_status yellow
 hibernate.search.backends.backendName.index_defaults.refresh_after_write true
